feat(Rate): Use in app review and move positon of rating entrance
