+++ app/build.gradle
@@ -68,6 +68,9 @@ dependencies {
     kapt 'androidx.room:room-compiler:2.2.5'
     androidTestImplementation 'androidx.room:room-testing:2.2.5'
 
+    //play-related
+    implementation 'com.google.android.play:core:1.8.2'
+    implementation 'com.google.android.play:core-ktx:1.8.1'
 
     //useful-library
     implementation 'com.blankj:utilcodex:1.24.7'
