+++ app/src/main/java/com/symeonchen/wakeupscreen/pages/AboutThisPageActivity.kt
@@ -1,9 +1,6 @@
 package com.symeonchen.wakeupscreen.pages
 
 import android.app.AlertDialog
-import android.content.ActivityNotFoundException
-import android.content.Intent
-import android.net.Uri
 import android.os.Bundle
 import android.view.View
 import androidx.lifecycle.Observer
@@ -61,29 +58,6 @@ class AboutThisPageActivity : ScBaseActivity() {
 
             }, 5000)
         }
-
-
-        item_setting_give_star.setOnClickListener {
-            try {
-
-                val intent = Intent(Intent.ACTION_VIEW).apply {
-                    data = Uri.parse(
-                        "https://play.google.com/store/apps/details?id=com.symeonchen.wakeupscreen"
-                    )
-                    setPackage("com.android.vending")
-                }
-                startActivity(intent)
-            } catch (anfe: ActivityNotFoundException) {
-                val i = Intent(Intent.ACTION_VIEW)
-                i.data = Uri.parse(
-                    "https://play.google.com/store/apps/details?id=com.symeonchen.wakeupscreen"
-                )
-                startActivity(i)
-            } catch (e: Exception) {
-                e.printStackTrace()
-            }
-        }
-
 
         settingModel.switchOfDebugMode.observe(this, Observer {
             item_setting_debug_mode_toast.bindData(
