+++ app/src/main/res/layout/activity_about_this.xml
@@ -63,13 +63,6 @@
                 app:sc_setting_item_title_main="@string/app_introduce"
                 app:sc_setting_single_line="true" />
 
-            <com.symeonchen.uicomponent.views.SCSettingItem
-                android:id="@+id/item_setting_give_star"
-                android:layout_width="match_parent"
-                android:layout_height="70dp"
-                app:sc_setting_item_title_main="@string/give_star"
-                app:sc_setting_item_title_second="@string/click_here_to_open_google_play_store" />
-
             <TextView
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
